requirements.txt
spotipy==2.23.0
python-dotenv==1.0.0
yt-dlp==2023.7.6
youtube-search==2.1.2
youtube-search-python==1.6.6